CodeSnip version 3.13.0

+++ Src/FmDeleteCategoryDlg.pas
@@ -1,15 +1,36 @@
 {
- * This Source Code Form is subject to the terms of the Mozilla Public License,
- * v. 2.0. If a copy of the MPL was not distributed with this file, You can
- * obtain one at http://mozilla.org/MPL/2.0/
+ * FmDeleteCategoryDlg.pas
  *
- * Copyright (C) 2009-2013, Peter Johnson (www.delphidabbler.com).
+ * Implements a dialog box that permits user to select and delete a user-defined
+ * category.
  *
  * $Rev$
  * $Date$
  *
- * Implements a dialogue box that permits user to select and delete a user
- * defined category.
+ * ***** BEGIN LICENSE BLOCK *****
+ *
+ * Version: MPL 1.1
+ *
+ * The contents of this file are subject to the Mozilla Public License Version
+ * 1.1 (the "License"); you may not use this file except in compliance with the
+ * License. You may obtain a copy of the License at http://www.mozilla.org/MPL/
+ *
+ * Software distributed under the License is distributed on an "AS IS" basis,
+ * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License for
+ * the specific language governing rights and limitations under the License.
+ *
+ * The Original Code is FmDeleteCategoryDlg.pas
+ *
+ * The Initial Developer of the Original Code is Peter Johnson
+ * (http://www.delphidabbler.com/).
+ *
+ * Portions created by the Initial Developer are Copyright (C) 2009 Peter
+ * Johnson. All Rights Reserved.
+ *
+ * Contributor(s)
+ *   NONE
+ *
+ * ***** END LICENSE BLOCK *****
 }
 
 
@@ -23,7 +44,7 @@ uses
   // Delphi
   Forms, StdCtrls, Controls, ExtCtrls, Classes,
   // Project
-  DB.UCategory, FmCategoryEditDlg, FrCategoryList, UBaseObjects;
+  FmCategoryEditDlg, FrCategoryList, UBaseObjects, USnippets;
 
 
 type
@@ -33,7 +54,6 @@ type
   }
   TDeleteCategoryDlg = class(TCategoryEditDlg, INoPublicConstruct)
     frmCategories: TCategoryListFrame;
-    lblErrorMsg: TLabel;
     procedure btnOKClick(Sender: TObject);
   strict private
     fCategories: TCategoryList; // List of categories that can be deleted
@@ -41,10 +61,6 @@ type
       {Handles category list frame's change event. Updates state of OK button
       according to changes.
         @param Sender [in] Not used.
-      }
-    procedure UpdateErrorLabelState;
-      {Shows or hides error state label depending of whether selected category
-      can be deleted.
       }
     procedure DeleteCategory(const Cat: TCategory);
       {Deletes category and all its snippets from database.
@@ -76,8 +92,10 @@ implementation
 
 
 uses
+  // Delphi
+  SysUtils,
   // Project
-  DB.UMain, UColours, UCtrlArranger, UFontHelper;
+  UConsts, UMessageBox;
 
 {$R *.dfm}
 
@@ -89,8 +107,6 @@ procedure TDeleteCategoryDlg.ArrangeForm;
   }
 begin
   frmCategories.ArrangeFrame;
-  TCtrlArranger.SetLabelHeight(lblErrorMsg);
-  lblErrorMsg.Top := TCtrlArranger.BottomOf(frmCategories, 8);
   inherited;
 end;
 
@@ -101,28 +117,37 @@ procedure TDeleteCategoryDlg.btnOKClick(Sender: TObject);
   }
 var
   Cat: TCategory; // selected category
+resourcestring
+  // Confirmation prompt
+  sConfirmDelete = 'The "%0:2s" category contains %1:d snippet(s).'
+    + EOL2
+    + 'If you delete the category all the snippet(s) will also be deleted.'
+    + EOL2 + EOL
+    + 'Are you sure you want to delete "%0:2s"?';
 begin
   inherited;
   Cat := frmCategories.SelectedCategory;
   Assert(Assigned(Cat), ClassName + '.btnOKClick: No category selected');
-  Assert(Cat.CanDelete, ClassName + '.btnOKClick: Category can''t be deleted');
-  DeleteCategory(Cat);
+  if (Cat.Routines.Count = 0)
+    or TMessageBox.Confirm(
+      Self, Format(sConfirmDelete, [Cat.Description, Cat.Routines.Count])
+    ) then
+    DeleteCategory(Cat)
+  else
+    ModalResult := mrNone;
 end;
 
 procedure TDeleteCategoryDlg.ConfigForm;
-  {Configures form. Sets up controls and supplies event handler to frame.
+  {Configures form. Populates controls and supplies event handler to frame.
   }
 resourcestring
   // Prompt text for frame
-  sPrompt = 'Select &category to be deleted:';
+  sPrompt = 'Select category to be deleted:';
 begin
   inherited;
   frmCategories.OnChange := SelectionChangeHandler;
   frmCategories.Prompt := sPrompt;
   frmCategories.SetCategories(fCategories);
-  TFontHelper.SetDefaultFont(lblErrorMsg.Font);
-  lblErrorMsg.Font.Color := clWarningText;
-  lblErrorMsg.Visible := False;
 end;
 
 procedure TDeleteCategoryDlg.DeleteCategory(const Cat: TCategory);
@@ -130,7 +155,7 @@ procedure TDeleteCategoryDlg.DeleteCategory(const Cat: TCategory);
     @param Cat [in] Category to be deleted.
   }
 begin
-  (Database as IDatabaseEdit).DeleteCategory(Cat);
+  (Snippets as ISnippetsEdit).DeleteCategory(Cat);
 end;
 
 class function TDeleteCategoryDlg.Execute(AOwner: TComponent;
@@ -157,18 +182,6 @@ procedure TDeleteCategoryDlg.SelectionChangeHandler(Sender: TObject);
   }
 begin
   UpdateOKBtn;
-  UpdateErrorLabelState;
-end;
-
-procedure TDeleteCategoryDlg.UpdateErrorLabelState;
-  {Shows or hides error state label depending of whether selected category can
-  be deleted.
-  }
-begin
-  if not frmCategories.IsValidEntry then
-    lblErrorMsg.Visible := False
-  else
-    lblErrorMsg.Visible := not frmCategories.SelectedCategory.CanDelete;
 end;
 
 procedure TDeleteCategoryDlg.UpdateOKBtn;
@@ -176,8 +189,7 @@ procedure TDeleteCategoryDlg.UpdateOKBtn;
   dialog box.
   }
 begin
-  btnOK.Enabled := frmCategories.IsValidEntry
-    and frmCategories.SelectedCategory.CanDelete;
+  btnOK.Enabled := frmCategories.IsValidEntry;
 end;
 
 end.
